@@ -22,7 +22,7 @@
 		"ml-regression-simple-linear": "3.0.1",
 		"regression": "^2.0.1",
 		"sharp": "^0.33.5",
-		"svelte": "^4.2.19",
+		"svelte": "^5.1.4",
 		"typescript": "^5.6.3",
 		"zod": "^3.23.8"
 	}
