@@ -1,5 +1,9 @@
 <script lang="ts">
-	export let repo: string;
+	interface Props {
+		repo: string;
+	}
+
+	let { repo }: Props = $props();
 </script>
 
 <a href={'https://github.com/' + repo} target="_blank">{repo}</a>
