@@ -1,8 +1,12 @@
 <script lang="ts">
 	import slug from 'slug';
 
-	export let id: string;
-	export let name: string;
+	interface Props {
+		id: string;
+		name: string;
+	}
+
+	let { id, name }: Props = $props();
 </script>
 
 <a href={'/obsidian-stats/themes/' + slug(id)}>{name}</a>
